+++ packages/uniforms-semantic/src/SelectField.js
@@ -42,6 +42,7 @@ const renderCheckboxes = ({
       </div>
     </div>
   ));
+
 const renderSelect = ({
   allowedValues,
   disabled,
@@ -60,13 +61,15 @@ const renderSelect = ({
     disabled={disabled}
     id={id}
     name={name}
-    onChange={event => onChange(event.target.value)}
+    onChange={event =>
+      onChange(event.target.value !== '' ? event.target.value : undefined)
+    }
     ref={inputRef}
     value={value}
   >
-    {(!!placeholder || !required) && (
+    {(!!placeholder || !required || value === '') && (
       <option value="" disabled={required} hidden={required}>
-        {placeholder ? placeholder : label}
+        {placeholder || label}
       </option>
     )}
 
@@ -77,6 +80,7 @@ const renderSelect = ({
     ))}
   </select>
 );
+
 const Select = ({
   allowedValues,
   checkboxes,
@@ -134,4 +138,5 @@ const Select = ({
     )}
   </div>
 );
+
 export default connectField(Select);